refactor:  Employee 관련 컴포넌트 리팩토링
회의 결론에 의거 리팩토링을 진행하였음.
