+++ src/main/java/dau/azit/simon/employee/controller/EmployeeController.java
@@ -1,10 +1,11 @@
 package dau.azit.simon.employee.controller;
 
 import dau.azit.simon.config.SimonConfigProperties;
-import dau.azit.simon.employee.dto.EmployeeCreateRequestDto;
-import dau.azit.simon.employee.dto.EmployeeLoginRequestDto;
-import dau.azit.simon.employee.dto.EmployeeUpdateRequestDto;
-import dau.azit.simon.employee.entity.Employee;
+import dau.azit.simon.employee.dto.request.EmployeeCreateDto;
+import dau.azit.simon.employee.dto.request.EmployeeLoginDto;
+import dau.azit.simon.employee.dto.request.EmployeeUpdateDto;
+import dau.azit.simon.employee.dto.response.EmployeeCheckDto;
+import dau.azit.simon.employee.domain.Employee;
 import dau.azit.simon.employee.service.EmployeeService;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
@@ -27,32 +28,35 @@ public class EmployeeController {
     private final EmployeeService employeeService;
 
     @GetMapping
-    public ResponseEntity<Page<Employee>> employeeList(Pageable pageable) {
-        return ResponseEntity.ok(employeeService.findAllEmployee(pageable));
+    public ResponseEntity<Page<EmployeeCheckDto>> employeeList(Pageable pageable) {
+        Page<Employee> employeePage = employeeService.findAllEmployee(pageable);
+        return ResponseEntity.ok(employeePage.map(EmployeeCheckDto::from));
     }
 
     @GetMapping("/{uid}")
-    public ResponseEntity<Employee> employeeDetails(@PathVariable UUID uid) {
-        return ResponseEntity.ok(employeeService.findOneByUid(uid));
+    public ResponseEntity<EmployeeCheckDto> employeeDetails(@PathVariable UUID uid) {
+        Employee employee = employeeService.findOneByUid(uid);
+        return ResponseEntity.ok(EmployeeCheckDto.from(employee));
     }
 
     @PostMapping
-    public ResponseEntity<Employee> employeeRegister(@RequestBody EmployeeCreateRequestDto dto) {
+    public ResponseEntity<EmployeeCheckDto> employeeRegister(@RequestBody EmployeeCreateDto dto) {
         Employee employee = employeeService.register(dto);
         String currentUri = ServletUriComponentsBuilder.fromCurrentRequestUri().toUriString();
-        return ResponseEntity.created(URI.create(currentUri + "/" + employee.getUid())).body(employee);
+        return ResponseEntity.created(URI.create(currentUri + "/" + employee.getUid())).body(EmployeeCheckDto.from(employee));
     }
 
     @PostMapping("/login")
-    public ResponseEntity<String> login(HttpSession session, @RequestBody EmployeeLoginRequestDto dto) {
+    public ResponseEntity<String> login(HttpSession session, @RequestBody EmployeeLoginDto dto) {
         employeeService.login(dto);
         session.setAttribute(properties.getKey(), employeeService.findOneByPhoneNumber(dto.phoneNumber()));
         return ResponseEntity.ok("Authentication Success");
     }
 
     @PutMapping("/{uid}")
-    public ResponseEntity<Employee> employeeModify(@PathVariable UUID uid, @RequestBody EmployeeUpdateRequestDto dto) {
-        return ResponseEntity.ok(employeeService.modifyOne(uid, dto));
+    public ResponseEntity<EmployeeCheckDto> employeeModify(@PathVariable UUID uid, @RequestBody EmployeeUpdateDto dto) {
+        Employee employee = employeeService.modifyOne(uid, dto);
+        return ResponseEntity.ok(EmployeeCheckDto.from(employee));
     }
 
     @DeleteMapping("/{uid}")
