+++ src/main/java/dau/azit/simon/employee/domain/UserRole.java
@@ -1,0 +1,5 @@
+package dau.azit.simon.employee.domain;
+
+public enum UserRole {
+    MANAGER, ADMIN
+}
