--- src/main/java/dau/azit/simon/employee/dto/EmployeeCreateRequestDto.java
@@ -1,10 +1,0 @@
-package dau.azit.simon.employee.dto;
-
-public record EmployeeCreateRequestDto(
-        String role,
-        String name,
-        String address,
-        String description,
-        String phoneNumber,
-        String password
-) {}
