--- src/main/java/dau/azit/simon/employee/dto/EmployeeLoginRequestDto.java
@@ -1,7 +1,0 @@
-package dau.azit.simon.employee.dto;
-
-public record EmployeeLoginRequestDto(
-        String phoneNumber,
-        String password
-) {
-}
