--- src/main/java/dau/azit/simon/employee/dto/EmployeeUpdateRequestDto.java
@@ -1,7 +1,0 @@
-package dau.azit.simon.employee.dto;
-
-public record EmployeeUpdateRequestDto(
-        String name,
-        String address,
-        String description
-) {}
