+++ src/main/java/dau/azit/simon/employee/dto/request/EmployeeCreateDto.java
@@ -1,0 +1,10 @@
+package dau.azit.simon.employee.dto.request;
+
+public record EmployeeCreateDto(
+        String role,
+        String name,
+        String address,
+        String description,
+        String phoneNumber,
+        String password
+) {}
