+++ src/main/java/dau/azit/simon/employee/dto/request/EmployeeLoginDto.java
@@ -1,0 +1,7 @@
+package dau.azit.simon.employee.dto.request;
+
+public record EmployeeLoginDto(
+        String phoneNumber,
+        String password
+) {
+}
