+++ src/main/java/dau/azit/simon/employee/dto/request/EmployeeUpdateDto.java
@@ -1,0 +1,7 @@
+package dau.azit.simon.employee.dto.request;
+
+public record EmployeeUpdateDto(
+        String name,
+        String address,
+        String description
+) {}
