--- src/main/java/dau/azit/simon/employee/entity/UserDetailsImpl.java
@@ -1,53 +1,0 @@
-package dau.azit.simon.employee.entity;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.Collection;
-import java.util.List;
-
-@RequiredArgsConstructor
-public class UserDetailsImpl implements UserDetails {
-
-    @Getter
-    private final Employee employee;
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return List.of(new SimpleGrantedAuthority(employee.getRole().name()));
-    }
-
-    @Override
-    public String getPassword() {
-        return employee.getPassword();
-    }
-
-    @Override
-    public String getUsername() {
-        return employee.getPhoneNumber();
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-
-}
