--- src/main/java/dau/azit/simon/employee/entity/UserRole.java
@@ -1,5 +1,0 @@
-package dau.azit.simon.employee.entity;
-
-public enum UserRole {
-    MANAGER, ADMIN
-}
