+++ src/main/resources/application.yml
@@ -24,5 +24,5 @@ server:
       cookie:
         name: _simon_session
 
-
-
+session:
+  key: ${SIMON_SESSION_KEY}
