--- src/main/resources/simon.yml
@@ -1,2 +1,0 @@
-session:
-  key: SIMON_SESSION_KEY
